refactor(#1): tanstack query 설정 분리 및 구조 개선
