+++ src/lib/tanstack/getQueryClient.ts
@@ -1,0 +1,15 @@
+import { isServer, QueryClient } from "@tanstack/react-query";
+import { makeQueryClient } from "./makeQueryClient";
+
+let browserQueryClient: QueryClient | undefined = undefined;
+
+export default function getQueryClient() {
+  if (isServer) {
+    return makeQueryClient();
+  } else {
+    if (!browserQueryClient) {
+      browserQueryClient = makeQueryClient();
+    }
+    return browserQueryClient;
+  }
+}
